+++ sentry_link/lib/sentry_link.dart
@@ -1,3 +1,4 @@
 library sentry_link;
 
 export 'src/sentry_link.dart' show SentryLink;
+export 'src/gql_event_processor.dart';