+++ sentry_link/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+- Fix an invalid usage of Sentry's context
+- Add event processor for nested LinkExceptions
+
 ## 0.0.2
 
 - Update dependencies and add some docs